Add algorithms dependency on commons-math 3.0
Note that since 3.0 is not released yet, this a SNAPSHOT dependency on
the latest 3.0 build. So builds may be less stable until the final 3.0
is released and we can switch to it. Also, if we decide to do a release
version of imglib2 soon, we will need to create an interim release
version of commons-math-3.0 on which it would depend, so that the
imglib2 release build is totally stable.