changed equalIterationOrder() mechanism.
Now every IterableRealInterval has a method iterationOrder() which
returns an Object representing its iteration order.  Comparison of
iteration orders is via Object.equals().  That is, if you want to
find whether IterableIntervals A and B have equal iteration order,
you ask whether A.iterationOrder().equals(B.iterationOrder()).

equalIterationOrder() is deprecated. I still changed it everywhere
to have it still work but some day we should remove it to keep the
IterableRealInterval interface as small as possible.