Redesign how histograms are counted
The desire to allow users to add and subtract subregions of data from an
existing histogram has initiated a change in how histograms populate
themselves. It is no longer the case that one must provide a data source to
the constructors. Rather one uses countData() providing a data source to be
used to populate the histogram counts. Then one can addData() and
subtractData() as desired.

The only loss of functionality here is the recalc() method is now gone. To
support it in the current scheme would have required some way of tracking
adds and subtracts internally. This is both complex and wasteful of
resources.

Note there still are constructors that take an initial data source for the
simple use cases.