Update to imglib2 2.0.1
This updates the POM to use the standalone pom-imglib2 parent.
We no longer need to declare most versions, since they are managed.

The artifact was renamed to imglib2-algorithm (from imglib2-algorithms)
for consistency with its package prefix. Because of that, we can start
the versioning over at 0.1.0, which according to SemVer indicates the
initial incubation period of development.

This commit also updates the code as necessary to address backwards
incompatible changes to the ImgLib2 library in the wake of the release.