Add deprecated net.imglib2.multithreading package
This code was removed from imglib2 core for the 2.0.1 release, but is
still in use by imglib2-algorithm.

As discussed at the October 2014 Madison ImgLib2 hackathon, the eventual
plan is to improve scijava-common's org.scijava.thread package to have
baseline multithreading interface(s), which would be used here in
imglib2-algorithm, in imagej-ops, and elsewhere as appropriate for
supporting multithreading computation.

First, though, we want to study the new constructs of Java 8, such as
java.util.Spliterator, as well as third party libraries
(quartz-scheduler.org, perhaps) and Scala (scala.concurrent package).