Fix another problem with border condition and chaining strels for Erosion.
The problem mentioned in the previous commits also affects the
"toTarget" method. Depending on whether the user is using a decomposed
strel or a non-decomposed strel, results could be completely different
for some border case (typically when an interesting pixel was off the
target interval, but could be reached by the strel when iterating on the
border of the interval).

To fix this, like previsouly, use the full erosion before copying the
results back to target, which is not that easy to come up with.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>