Iterates over a Bresenham circle in the target RandomAccessible.
Each point of the circle is iterated exactly once, and
there is no "hole" in the circle. Contrary to the algorithm
in Wikipedia, the circles generated by this cursor are "slim":
each pixel of the circle is connected with 8-connectivity.
For instance, a bitmap excerpt from such a circle looks like this:

 ..........
 OOO.......
 ...OO.....
 .....O....
 .....O....