Use MathJax for math formulas
This is much more elegant and succinct.

It does, however, create an impeding problem: javadoc 1.8.0_121 and
later requires passing a new --allow-script-in-javadoc flag, or else the
javadoc generation will fail due to the new header using a <script> tag.

Right now, Travis CI uses 1.8.0_111, so the build will pass... but we
will have to figure out what to do once Travis updates to a newer Java.