HoughLineTransform: Clone peak storage array
This stage of the transform makes use of a single array to grab each
peak Point and then store it in an ArrayList of long[]. However the
array was not copied as it was added into the ArrayList, resulting in
all entries in the ArrayList pointing to the same array. This commit
fixes this bug through use of Arrays.copyOf() to copy each array and add
the result to the ArrayList.