VoteLines: fix predicate bug
This commit changes the Predicate deciding whether to allow votespace
population from being a strictly greater-than to a
greater-than-or-equal-to comparison. This fix was made to prevent a bug
in which no threshold is provided to the algorithm and at the same time
the max value of the image is at the first index of the image. If the
comparison is not <= then all pixels will fail the Predicate, since no
pixel is greater than the max. This means that passing a value to the
algorithm is still preferable but allows more scenarios where no value
is given.