Rename LoopMath to ImgMath, and switch to: new ImgMath( operation ).into( target )
The new approach makes possible the caching (in a script or program)
of all the preliminary setup of the operations, and then merely
apply them repeatedly if desired.

Also reads nicer, separating concerns: the operation in the constructor,
and the target image into which to apply the result of the operation in
an explicit method.