ImgMath: implemented IF, EQ, NEQ, LT, GT
Equal (EQ), NotEqual (NEQ), LessThan (LT), and GreaterThan (GT)
all set the output to 1 when true, to zero when false.

IF is a TrinaryFunction that conditionally executes either the 2nd ("then")
or the 3rd ("else") argument depending on whether the 1st is true or false.

Performance tests show a considerable cost, perhaps 3X over low-level math.