Make a AStar algorithm in 2D for ImgLib2, using an image as weight.
This implementation is based on the garbage-collection free version
of Wang Yuntao. It therefore has the same limitations:
- Only 2D.
- Image must not be too big (max 2^16 in width and height).

The reason is that W.Y. used a sngle long value to store all the node
values (x, y, gn and fn with bit shifts).

Te original code has been heavility modified so that it can:
- Accept an ImgLib2 image that will serve as weights during the
pathfinding.
- Pixels with intensities that are below a specified threshold are
never visited.
- A configurable penalty allows penalizing pixels with low intensities.
Right now the penalty function is a linear function between the max
and the min thresholded intensity of the source image, with an
amplitude that can be specified.
- The heuristics can be set arbitrarily. By default we have the Chebyshev
distance. Or the Euclidean distance.
- The directions in which pixels are visited can be configured. For
instance you can forbid going up and left.